third_party/blink/web_tests/fast/forms/select/select-add-expected.txt

Test select.add() method

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test select
PASS addOption(0) is "X,0,1,2"
PASS addOption(1) is "0,X,1,2"
PASS addOption(2) is "0,1,X,2"
PASS addOption(3) is "0,1,2,X"
PASS addOption(100) is "0,1,2,X"
PASS addOption(-100) is "0,1,2,X"
PASS getSelectOptions() is "0,1,2,X"
PASS addOption(undefined) is "0,1,2,X"
PASS addOption(null) is "0,1,2,X"
PASS addOption(NaN) is "X,0,1,2"
PASS addOption(Infinity) is "X,0,1,2"
PASS addOption(-Infinity) is "X,0,1,2"
PASS addOption("foo") is "X,0,1,2"
PASS addOption("0") is "X,0,1,2"
PASS addOption("1") is "0,X,1,2"
PASS addOption(false) is "X,0,1,2"
PASS addOption(true) is "0,X,1,2"
PASS addOption(2147483647) is "0,1,2,X"
PASS addOption(-2147483647) is "0,1,2,X"
PASS addOption(2147483648) is "0,1,2,X"
PASS addOption(-2147483648) is "0,1,2,X"
PASS select.add() threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': 1 argument required, but only 0 present..
PASS select.add("foo") threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add("foo", 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(undefined) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(undefined, 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(null) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(null, 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(new Option("X", "X", false, false), 0, 1) is undefined.
PASS select.add(new Option("X", "X", false, false), select.options[0], 2) is undefined.
PASS getSelectOptions() is "0,X,Y,1,2"
Test select multiple
PASS addOption(0) is "X,0,1,2"
PASS addOption(1) is "0,X,1,2"
PASS addOption(2) is "0,1,X,2"
PASS addOption(3) is "0,1,2,X"
PASS addOption(100) is "0,1,2,X"
PASS addOption(-100) is "0,1,2,X"
PASS getSelectOptions() is "0,1,2,X"
PASS addOption(undefined) is "0,1,2,X"
PASS addOption(null) is "0,1,2,X"
PASS addOption(NaN) is "X,0,1,2"
PASS addOption(Infinity) is "X,0,1,2"
PASS addOption(-Infinity) is "X,0,1,2"
PASS addOption("foo") is "X,0,1,2"
PASS addOption("0") is "X,0,1,2"
PASS addOption("1") is "0,X,1,2"
PASS addOption(false) is "X,0,1,2"
PASS addOption(true) is "0,X,1,2"
PASS addOption(2147483647) is "0,1,2,X"
PASS addOption(-2147483647) is "0,1,2,X"
PASS addOption(2147483648) is "0,1,2,X"
PASS addOption(-2147483648) is "0,1,2,X"
PASS select.add() threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': 1 argument required, but only 0 present..
PASS select.add("foo") threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add("foo", 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(undefined) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(undefined, 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(null) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(null, 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS select.add(new Option("X", "X", false, false), 0, 1) is undefined.
PASS select.add(new Option("X", "X", false, false), select.options[0], 2) is undefined.
PASS getSelectOptions() is "0,X,Y,1,2"
PASS successfullyParsed is true

TEST COMPLETE
